refactor into a hook
